// apps/admin/src/api/generated/model/listFeedbackCampaignResultsParams.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Join The Six API
 * Operations API for Join The Six
 * OpenAPI spec version: 1.0.0
 */
import type { ListFeedbackCampaignResultsQuestionKey } from "./listFeedbackCampaignResultsQuestionKey";
import type { ListFeedbackCampaignResultsReviewStatus } from "./listFeedbackCampaignResultsReviewStatus";

export type ListFeedbackCampaignResultsParams = {
  questionKey?: ListFeedbackCampaignResultsQuestionKey;
  /**
   * @pattern ^([0-9a-fA-F]{8}-[0-9a-fA-F]{4}-[1-8][0-9a-fA-F]{3}-[89abAB][0-9a-fA-F]{3}-[0-9a-fA-F]{12}|00000000-0000-0000-0000-000000000000|ffffffff-ffff-ffff-ffff-ffffffffffff)$
   */
  participantId?: string;
  reviewStatus?: ListFeedbackCampaignResultsReviewStatus;
};