// apps/admin/src/api/generated/model/listFeedbackCampaignResultsReviewStatus.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Join The Six API
 * Operations API for Join The Six
 * OpenAPI spec version: 1.0.0
 */

export type ListFeedbackCampaignResultsReviewStatus =
  (typeof ListFeedbackCampaignResultsReviewStatus)[keyof typeof ListFeedbackCampaignResultsReviewStatus];

export const ListFeedbackCampaignResultsReviewStatus = {
  new: "new",
  dismissed: "dismissed",
} as const;